Fix/add duplicate author data
Because of a misprioritized if statement in gather_persons.py, people
who had duplicate names were all assigned the affiliation and sector
based on the email address of the survey author of one of them.
In addition to fixing this bug, I went over all the duplicates that were
missing a GS profile and tried to fit them with one with newer GS data.