wrote code to create a data frame where each row is a unique author name with the number of unique coauthors who are male and female respectively that they have written a paper with.
I added this to the repo as `data/nonsys_collaborator_counts` because it takes a long time to create the dataframe so it seems easier to just read it in after we've created it once.

Out of this data frame came two tables: the first shows the average number of collaborators broken down by gender. the second shows the average number of male collaborators and average number of female collaborators also broken down by gender.