Correct productivity metrics in collab.Rmd
The main problem was the merging of system and non-system data, which
meant that some authors had separate counts of `as_authors`, even if
they're the same person. Now productivity is computed by name+gender
only. This leads to an opposite problem with names that haven't been
disambiguated between sys and nonsys (to be addressed in a later
commit).